Base completa del backend de la app

+++ src/main/java/com/gestion/infrastructure/persistence/repositories/DocumentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentRepository extends JpaRepository<DocumentDAO, Integer> {
+public interface DocumentRepository extends JpaRepository<DocumentDAO, Long> {
 }
